MERN layout
